requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
ast2json==0.4
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
autobahn==24.4.2
Automat==24.8.1
Babel==2.15.0
beautifulsoup4==4.12.3
big_O==0.11.0
bleach==6.1.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.16.0
channels==3.0.5
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
daphne==3.0.2
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.31.0
distlib==0.3.8
distro==1.9.0
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
django-jet==1.0.8
django-jet-reboot==1.3.10
django-modelcluster==6.4
django-permissionedforms==0.1
django-taggit==6.1.0
django-treebeard==4.7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
draftjs-exporter==5.0.0
entrypoints==0.4
et_xmlfile==2.0.0
executing==2.0.1
faiss-cpu==1.9.0
Faker==30.8.0
fastjsonschema==2.20.0
favicon==0.7.0
filelock==3.15.4
filetype==1.2.0
Flask==3.1.1
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.26.2
hyperlink==21.0.0
idna==3.7
importlib_metadata==8.5.0
incremental==24.7.2
ipykernel==6.29.5
ipython==8.26.0
ipython-genutils==0.2.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kiwisolver==1.4.7
l18n==2021.3
labelImg==1.8.6
laces==0.1.2
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.12
langchain-ollama==0.1.1
langchain-text-splitters==0.3.0
langsmith==0.1.136
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
marshmallow==3.23.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.0.5
mutagen==1.46.0
myenv==201902.7
mypy-extensions==1.0.0
mysql==0.0.3
mysqlclient==2.2.4
narwhals==1.9.4
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
ollama==0.3.1
openai==0.28.0
openpyxl==3.1.5
orjson==3.10.7
overrides==7.7.0
packaging==24.1
panda==0.3.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pillow==10.4.0
pillow_heif==0.21.0
pipenv==2024.0.3
platformdirs==4.2.2
plotly==5.24.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.28.2
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.8.2
pydantic-settings==2.6.0
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pymdown-extensions==10.11.2
PyMySQL==1.1.1
pyOpenSSL==24.2.1
pyparsing==3.2.0
PyPDF2==3.0.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
redis==6.1.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.2
rpds-py==0.19.1
safetensors==0.4.5
Send2Trash==1.8.3
service-identity==24.2.0
setuptools==72.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spotipy==2.23.0
SQLAlchemy==2.0.32
sqlparse==0.5.1
st-annotated-text==4.0.1
st-theme==1.2.3
stack-data==0.6.3
streamlit==1.39.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.5.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.13.2
telepath==0.3.1
tenacity==8.5.0
terminado==0.18.1
tiktoken==0.8.0
tinycss2==1.3.0
tokenizers==0.20.1
toml==0.10.2
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.46.1
Twisted==24.10.0
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
validators==0.34.0
vboxapi==1.0
virtualenv==20.26.3
virtualenvwrapper-win==1.2.7
wagtail==6.3.2
watchdog==5.0.3
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
whitenoise==6.7.0
widgetsnbextension==4.0.11
Willow==1.9.0
yarl==1.9.4
zipp==3.20.2
zope.interface==7.1.1
